add mapping for aggregations in text editor and summary cards
