+++ libs/shared/src/lib/components/widgets/common/aggregation-settings/aggregation-settings.component.html
@@ -74,7 +74,7 @@
       <h3 class="mb-1">{{ 'components.mapping.title' | translate }}</h3>
       <shared-series-mapping
         [availableFields]="availableSeriesFields"
-        [formGroup]="$any(formGroup.get('chart.mapping'))"
+        [formGroup]="$any(formGroup.get('aggregation.mapping'))"
       ></shared-series-mapping>
     </ng-container>
   </ng-container>
