+++ libs/shared/src/lib/services/aggregation-builder/aggregation-builder.service.ts
@@ -4,6 +4,9 @@ import { Accumulators } from '../../components/ui/aggregation-builder/pipeline/e
 import { BehaviorSubject, Observable } from 'rxjs';
 import { addNewField } from '../../components/query-builder/query-builder-forms';
 import { cloneDeep } from 'lodash';
+import { Aggregation } from '../../models/aggregation.model';
+import { Resource } from '../../models/resource.model';
+import { QueryBuilderService } from '../query-builder/query-builder.service';
 
 /**
  * Shared aggregation service.
@@ -15,6 +18,13 @@ import { cloneDeep } from 'lodash';
 })
 export class AggregationBuilderService {
   private gridSubject = new BehaviorSubject<any>(null);
+
+  /**
+   * Aggregation builder service
+   *
+   * @param queryBuilder shared query builder service
+   */
+  constructor(private queryBuilder: QueryBuilderService) {}
 
   /**
    * Get the data for grid preview as an observable.
@@ -194,5 +204,48 @@ export class AggregationBuilderService {
       outField = { ...outField };
     }
     return outField;
+  }
+
+  /**
+   * Set available series fields, from resource fields and aggregation definition.
+   *
+   * @param aggregation aggregation to get the series fields from
+   * @param resource resource
+   * @returns series fields
+   */
+  public getAvailableSeriesFields(
+    aggregation?: Aggregation,
+    resource?: Resource
+  ): any {
+    if (aggregation) {
+      const fields = this.queryBuilder
+        .getFields(resource?.queryName as string)
+        .filter(
+          (field: any) =>
+            !(
+              field.name.includes('_id') &&
+              (field.type.name === 'ID' ||
+                (field.type.kind === 'LIST' && field.type.ofType.name === 'ID'))
+            )
+        );
+      const selectedFields = aggregation.sourceFields
+        .map((x: string) => {
+          const field = fields.find((y) => x === y.name);
+          if (!field) return null;
+          if (field.type.kind !== 'SCALAR') {
+            Object.assign(field, {
+              fields: this.queryBuilder.deconfineFields(
+                field.type,
+                new Set().add(resource?.name).add(field.type.ofType?.name)
+              ),
+            });
+          }
+          return field;
+        })
+        .filter((x: any) => x !== null);
+      return this.fieldsAfter(selectedFields, aggregation?.pipeline);
+    } else {
+      return [];
+    }
   }
 }
